FEATURE: refactoring models

+++ services/recipes_services.py
@@ -1,13 +1,14 @@
 from repository import *
+from sqlalchemy.orm import joinedload
 from services.ingredients_services import IngredientServices
 from services.instructions_services import InstructionServices
-from utils.functions import highlight
+from werkzeug.exceptions import BadRequest, Forbidden, NotFound, Conflict
 
 
 class RecipeServices():
     """Handles recipe view business logic"""
     @staticmethod
-    def process_recipe_data(request, book_id):
+    def process_recipe_data(request, book_id, user_id):
         """Consolidate 'create recipe' process"""
         try:
             recipe = request["recipe"]
@@ -20,7 +21,7 @@ class RecipeServices():
 
         try:
             recipe_data = RecipeServices.process_recipe(
-                book_id=book_id, recipe_name=recipe["name"], notes=notes)
+                book_id=book_id, recipe_name=recipe["name"], notes=notes, user_id=user_id)
 
             if ingredients:
                 recipe_data["ingredients"] = RecipeServices.process_ingredients(
@@ -42,12 +43,12 @@ class RecipeServices():
             raise type(e)(f"Failed to process_recipe_data: {e}") from e
 
     @staticmethod
-    def process_recipe(book_id, recipe_name, notes):
+    def process_recipe(book_id, recipe_name, notes, user_id):
         """Adds recipes and associates new recipe to book"""
         try:
             recipe_data = RecipeRepo.create_recipe(
-                name=recipe_name, notes=notes)
-            
+                name=recipe_name, notes=notes, user_id=user_id)
+
             RecipeBookRepo.create_entry(
                 book_id=book_id, recipe_id=recipe_data["id"])
 
@@ -62,11 +63,11 @@ class RecipeServices():
         try:
             ingredients_data = IngredientServices.process_ingredient_components(
                 book_id=book_id, ingredients=ingredients)
-            
+
             if not ingredients_data:
                 raise ValueError(
                     f"No ingredients data returned for recipe {recipe_id}")
-            
+
             for ingredient in ingredients_data:
                 id = RecipeIngredientRepo.create_ingredient(
                     recipe_id=recipe_id,
@@ -85,11 +86,11 @@ class RecipeServices():
         """Processes consolidated instructions and associates each instruction to Recipe"""
         try:
             instructions_data = InstructionServices.process_instructions(
-                instructions=instructions,book_id=book_id)
+                instructions=instructions, book_id=book_id)
 
             if not instructions_data:
                 raise ValueError(
-                    f"No instructions data returned for book id: {book_id}") 
+                    f"No instructions data returned for book id: {book_id}")
 
             for instruction in instructions_data:
                 id = RecipeInstructionRepo.create_entry(
@@ -109,30 +110,51 @@ class RecipeServices():
             book = Book.query.get(book_id)
             if not book:
                 raise ValueError(f"No book found with ID {book_id}")
-
             recipes_instances = book.recipes
             for recipe_instance in recipes_instances:
                 recipe_build = {}
 
                 recipe = Recipe.serialize(recipe_instance)
-                
+
                 recipe_build.update(recipe)
 
                 instructions = InstructionServices.build_instructions(
                     instances=recipe_instance.instructions, recipe_id=recipe["id"])
                 recipe_build["instructions"] = instructions
 
-                ingredients = IngredientServices.build_ingredients(instance=recipe_instance)
+                ingredients = IngredientServices.build_ingredients(
+                    instance=recipe_instance)
                 recipe_build["ingredients"] = ingredients
 
                 complete_recipes.append(recipe_build)
             return complete_recipes
         except Exception as e:
             raise type(e)(f"RecipeServices - build_recipes error: {e}") from e
-
+    
+    @staticmethod 
+    def build_recipe(recipe_id):
+        """Builds individual recipes"""
+        try:
+            recipe = RecipeRepo.query_recipe(recipe_pk=recipe_id)
+            instructions = InstructionServices.build_instructions(instances=recipe.instructions, recipe_id=recipe_id)
+            ingredients = IngredientServices.build_ingredients(instance=recipe.ingredients)
+            return {
+                    "is_owned_by": recipe.created_by_id,
+                    "id":recipe.id, 
+                    "created_by_id":recipe.created_by_id, 
+                    "name": recipe.name,
+                    "ingredients": ingredients,
+                    "instructions": instructions,
+                    "notes": recipe.notes
+                    }
+        except Exception as e:
+            raise type(e)(f"RecipeServices - build_recipe error: {e}") from e
+        
     @staticmethod
-    def process_edit(data, recipe_id):
+    def process_edit(user_id, data, recipe_id):
         """Consolidates recipe edit process"""
+        if user_id is not data.get("created_by_id"):
+            raise Forbidden("Not authorized to make edits")
         try:
             name = data.get("name")
             ingredients = data.get("ingredients")
@@ -185,7 +207,8 @@ class RecipeServices():
                 amount = ingredient.get("amount")
                 unit = ingredient.get("unit")
                 if not item and not amount and not unit:
-                    raise ValueError("No values in ingredient components to edit") 
+                    raise ValueError(
+                        "No values in ingredient components to edit")
 
                 quantity_amount_id = amount["id"] if amount else None
                 quantity_unit_id = unit["id"] if unit else None
@@ -195,8 +218,9 @@ class RecipeServices():
                     recipe_ingredient = Ingredient.query.get(
                         ingredient["id"])
                     if not recipe_ingredient:
-                        raise NotFound(f"No ingredient matching id #: {ingredient['id']}")
-                    
+                        raise NotFound(
+                            f"No ingredient matching id #: {ingredient['id']}")
+
                     if amount:
                         recipe_ingredient.quantity_amount_id = quantity_amount_id
                     if unit:
@@ -213,7 +237,7 @@ class RecipeServices():
             raise type(e)(f"Failed to process_edit_ingredients: {e}") from e
 
     @staticmethod
-    def process_edit_instructions(instructions,recipe_id):
+    def process_edit_instructions(instructions, recipe_id):
         """Edits recipe's instructions by modifying RecipeInstruction association"""
         try:
             for instruction in instructions:
@@ -232,8 +256,35 @@ class RecipeServices():
             raise type(e)(f"Failed to process_edit_instructions: {e}") from e
 
     @staticmethod
-    def remove_recipe(recipe_id):
+    def share_recipe(auth_id, recipient, recipe_id):
+        """Process sharing user recipe with recipient"""
+        recipient = UserRepo.query_user_name(user_name=recipient)
+        recipe = RecipeRepo.query_recipe(recipe_pk=recipe_id)
+        recipe_build = RecipeServices.build_recipe(recipe_id=recipe_id)
+        if not recipient:
+            return {"message": "User not found", "error": "Not Found", "code": 404}
+        if auth_id == recipient.id:
+            return {"message": "Why are you sharing this with yourself???", 
+                    "error": "BadRequest", "code": 400}
+        if not recipe.is_owned_by(auth_id):
+            return {"message": "This is not yours to share...",
+                    "error": "Forbidden", "code": 403}
+        try:
+            message = RecipeRepo.create_recipe_link(
+                recipient_id=recipient.id, shared_id=recipe_id)
+            db.session.commit()
+            message["recipient_id"] = recipient.id
+            return {**message, "recipe": recipe_build}
+        except Exception as e:
+            db.session.rollback()
+            raise type(e)(
+                f"Failed to share_recipe error: {e}") from e
+
+    @staticmethod
+    def remove_recipe(auth_id, recipe_id, data):
         """Deletes book recipe"""
+        if auth_id is not int(data["createdById"]):
+            raise Forbidden("Not authorized to delete!")
         try:
             RecipeRepo.delete_recipe(recipe_id=recipe_id)
             db.session.commit()
@@ -241,4 +292,23 @@ class RecipeServices():
         except Exception as e:
             db.session.rollback()
             raise type(e)(
-                f"DeleteServices - remove_recipe error: {e}") from e
+                f"Failed to remove_recipe error: {e}") from e
+
+    @staticmethod
+    def remove_shared_recipe(authed_id, recipe_id, book_id):
+        """Verifies shared recipe belongs to user's shared book. Then deletes association"""
+        user_book = UserBookRepo.query_user_book(book_id=book_id, user_id=authed_id)
+        highlight(user_book,"!")
+        if not user_book:
+                raise NotFound("Not found")
+        is_book_type_shared = user_book.book.book_type 
+        if not is_book_type_shared: 
+            raise ForbiddenError("Forbidden request")
+        try:
+            highlight((book_id,recipe_id),"!")
+            response = RecipeBookRepo.remove_book_association(book_id=book_id,recipe_id=recipe_id)
+            db.session.commit()
+            return response
+        except Exception as e:
+            db.session.rollback()
+            raise type(e)(f"Failed to remove_shared_recipe error: {e}") from e
